authentication

+++ base/views.py
@@ -1,5 +1,8 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from django.http import HttpResponse
 from django.db.models import Q
 from .models import Room, Topic, User
 from django.contrib.auth import authenticate, login, logout
@@ -15,8 +18,13 @@ rooms = [
 
 def loginPage(request):
 
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    page = "login"
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -32,8 +40,28 @@ def loginPage(request):
         else:
             messages.error(request, 'Invalid username or password')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
+
+def registerPage(request):
+    page = "register"
+    form = UserCreationForm()
+    context = {'page': page, 'form': form}
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration')
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -46,11 +74,15 @@ def home(request):
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
+    room = Room.objects.get(id=int(pk))
+    messages = room.message_set.all()
     context = {
-        'room': Room.objects.get(id=int(pk))
+        'room': room,
+        'messages': messages
     }
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -61,9 +93,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here.")
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid:
@@ -72,8 +109,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here.")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
